Refactor searching and sorting traits
