+++ app/Livewire/Traits/WithSearching.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace App\Livewire\Traits;
+
+use Livewire\Attributes\Url;
+
+trait WithSearching
+{
+    #[Url]
+    public string $search = '';
+
+    public function updatedSearch(): void
+    {
+        $this->resetPage();
+    }
+}
