+++ app/Livewire/Traits/WithSorting.php
@@ -1,0 +1,23 @@
+<?php
+
+namespace App\Livewire\Traits;
+
+trait WithSorting
+{
+    public string $sortBy = 'name';
+
+    public string $sortDirection = 'asc';
+
+    public function sort(string $column): void
+    {
+        if ($this->sortBy === $column) {
+            $this->sortDirection = $this->sortDirection === 'asc' ? 'desc' : 'asc';
+        } else {
+            $this->sortBy = $column;
+            $this->sortDirection = 'asc';
+        }
+
+        $this->resetPage();
+    }
+
+}
